refactor(website): Interact
