+++ packages/website/src/pages/packages/[name]/[tag]/[variant]/interact/index.tsx
@@ -4,7 +4,7 @@ import { NextSeo } from 'next-seo';
 import defaultSEO from '@/constants/defaultSeo';
 import { usePackageNameTagVersionUrlParams } from '@/hooks/routing/usePackageVersionUrlParams';
 import { useCannonChains } from '@/providers/CannonProvidersProvider';
-import InteractComponent from '@/features/Packages/Interact';
+import InteractComponent from '@/features/Packages/interact/Interact';
 
 function generateMetadata({
   name,
